move mocks folder
