+++ src/poe-fetch/poe-fetch.service.spec.ts
@@ -1,13 +1,13 @@
 import { Test } from '@nestjs/testing';
 import { PoeFetchService } from './poe-fetch.service';
 import { HttpModule } from '@nestjs/axios';
-import poeHandlers, { poeFirstResponse, queryId } from '../mocks/handlers';
+import poeHandlers, { poeFirstResponse, queryId } from '../../mocks/handlers';
 import { setupServer } from 'msw/node';
 import { saveAnyJsonInFile } from '../tools/workingWithFile';
 import * as path from 'path';
 import * as fs from 'fs/promises';
-import { poeSecondResponse } from '../mocks/poeSecondResponse';
-import { mockPoeTradeDataItems } from '../mocks/mock-poe-trade-data-items';
+import { poeSecondResponse } from '../../mocks/poeSecondResponse';
+import { mockPoeTradeDataItems } from '../../mocks/mock-poe-trade-data-items';
 
 jest.mock('../tools/workingWithFile', () => {
   const originalModule = jest.requireActual('../tools/workingWithFile');
