+++ tsconfig.spec.json
@@ -1,5 +1,5 @@
 {
   "extends": "./tsconfig.json",
   "include": ["**/*.spec.ts"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "mocks"]
 }
